+++ src/widgets/posts/index.ts
@@ -1,1 +1,3 @@
+export { PostModalSSR } from './postView/PostModalSSR';
+export { PostsListSSR } from './postFeed/PostsListSSR';
 export { PostModalBrowser } from './postView/PostModalBrowser';
